pynear: add python package, remove rpc.py script
