+++ .gitattributes
@@ -1,4 +1,4 @@
 Cargo.lock linguist-generated=true -diff
 core/protos/src/autogenerated/* linguist-generated=true -diff
 nearlib/protos.js linguist-generated=true -diff
-scripts/protos/* linguist-generated=true -diff
+pynear/**/protos/* linguist-generated=true -diff
