+++ .gitlab-ci.yml
@@ -17,7 +17,7 @@ variables:
     wget -O /tmp/protoc.zip https://github.com/protocolbuffers/protobuf/releases/download/v3.6.1/protoc-3.6.1-linux-x86_64.zip &&
     unzip -p /tmp/protoc.zip bin/protoc | sudo tee /usr/local/bin/protoc 1> /dev/null &&
     sudo chmod 755 /usr/local/bin/protoc &&
-    sudo pip install protobuf
+    (cd pynear && python setup.py install)
 
 .cleanup_obsolete_cache: &cleanup_obsolete_cache
     find "${CACHE_ROOT}/target" -atime +2 -delete
