+++ pynear/src/near/__init__.py
@@ -1,0 +1,3 @@
+from pkgutil import extend_path
+
+__path__ = extend_path(__path__, __name__)
